Add eli like history option.
Added the customize variable 'slime-repl-eli-history-behavior' to
the slime-repl customize group.  This defaults to nil which does
not change the behavior of the history.  When set to t this enables
an eli like history function where the history is inserted at point
in the REPL instead of replacing all current input.