Support for grabbing multiple outputs in swank:eval-and-grab-output
It is often desirable to get output from time and trace in Emacs,
particularly in org-babel.  This patch extends
swank:eval-and-grab-output to support emitting *trace-output*; it also
introduces *error-output*, for good measure.

We check whether eval-and-grab-output caller presumes older interface
to it and return a legacy list of two elements instead of alist in
those cases.

This check can very likely be dropped when either Emacs 27 or Org 9.4
become unsupported.